Sort out some generators that have moved to scala-libs
