+++ pipeline/transformer/transformer_sierra/src/test/scala/weco/pipeline/transformer/sierra/transformers/SierraFormatTest.scala
@@ -3,8 +3,8 @@ package weco.pipeline.transformer.sierra.transformers
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import weco.catalogue.internal_model.work.Format.Books
-import weco.catalogue.source_model.generators.SierraDataGenerators
 import weco.catalogue.source_model.sierra.source.SierraMaterialType
+import weco.sierra.generators.SierraDataGenerators
 
 class SierraFormatTest
     extends AnyFunSpec
