+++ pipeline/transformer/transformer_sierra/src/test/scala/weco/pipeline/transformer/sierra/transformers/SierraSubjectsTest.scala
@@ -3,8 +3,8 @@ package weco.pipeline.transformer.sierra.transformers
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import weco.catalogue.internal_model.work.{Concept, Subject}
-import weco.catalogue.source_model.generators.SierraDataGenerators
 import weco.catalogue.source_model.sierra.marc.{MarcSubfield, VarField}
+import weco.sierra.generators.SierraDataGenerators
 
 class SierraSubjectsTest
     extends AnyFunSpec
